Глобальное обновление: -Была добавлена система распределения работников по правам. Есть три основные категории: -Администраторы - работники со всемии правами редактирования/изменения/удаления других пользователей. -Модератор - работники с правами редактирования/изменения других пользователей. -Пользователь - работники с правами просмотра таблиц производства. (В будущем будет добавлена функция формирования отчёта (например заказа или найма сотрудников за данный период) -Переработан GUI и разделение его для работников с разными правами -Соответсвенно в таблице employees было добавлено поле "Права" -К данному полю был создан фильтр в контроллере EmployeeController.java Изменения система авторизации: -Добавлен вывод на экран id (паспорт) сессии -Происходит проверка работника на наличие тех или иных прав, вследствие чего открывается то или иное окно с соответсвующим для прав функционалом. Недоделано: -Для модераторов не были созданы и протестированы GUI -Недобавлен функционал редактирования/изменения существующих пользователей -Неубран весь шаблонный код из уже существующих классов

+++ src/main/java/org/example/demotest/app_controllers/ClientAppController.java
@@ -11,8 +11,10 @@ import javafx.stage.Stage;
 import org.example.demotest.dto.ServiceRequestClient;
 import org.example.demotest.entities.Client;
 import org.example.demotest.entities.Reputation;
+import org.example.demotest.entities.Role;
 import org.example.demotest.managers.LoginManager;
 import org.example.demotest.services.ClientService;
+import org.example.demotest.services.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Controller;
@@ -20,13 +22,16 @@ import org.springframework.web.client.RestTemplate;
 
 import java.util.Arrays;
 
-import static org.example.demotest.app_controllers.LoginController.sessionID;
+import static org.example.demotest.app_controllers.EmployeeAppController.PhoneFilter;
 
 @Controller
 public class ClientAppController {
 
-    @Autowired
-    private ApplicationContext applicationContext;
+    private final ApplicationContext applicationContext;
+
+    public ClientAppController(ApplicationContext applicationContext) {
+        this.applicationContext = applicationContext;
+    }
 
     private TextFormatter<String> createAlphaFilter() {
         return new TextFormatter<>(change -> {
@@ -48,59 +53,38 @@ public class ClientAppController {
     }
 
     private TextFormatter<String> createPhoneFilter() {
-        return new TextFormatter<>(change -> {
-            String newText = change.getControlNewText();
-
-            if (newText.length() > 12) {
-                return null;
-            }
-
-            if (newText.startsWith("8")) {
-                newText = "+7" + newText.substring(1);
-                change.setText(newText);
-                change.setRange(0, change.getControlText().length());
-            }
-
-            if (newText.matches("(\\+7\\d*|8\\d*)")) {
-                return change;
-            }
-
-            return null;
-        });
+        return PhoneFilter();
     }
-    private RestTemplate restTemplate = new RestTemplate();
-    private String url = "http://localhost:8080/client-api/v1/clients";
+    private final RestTemplate restTemplate = new RestTemplate();
+    private final String url = "http://localhost:8080/client-api/v1/clients";
 
     //Таблица
     @FXML private TableView<Client> clientTable;
     @FXML private TableColumn<Client, Long> ClientId;
-    @FXML private TableColumn<Client, String> CompanyName;
-    @FXML private TableColumn<Client, String> ContactPerson;
-    @FXML private TableColumn<Client, String> PhoneNumber;
-    @FXML private TableColumn<Client, String> Email;
-    @FXML private TableColumn<Client, String> Address;
+    @FXML private TableColumn<Client, String> CompanyName, ContactPerson, PhoneNumber, Email, Address;
     @FXML private TableColumn<Client, Reputation> Reputation;
 
     //Поля для создания пользователя
-    @FXML private TextField companyName;
-    @FXML private TextField contactPerson;
-    @FXML private TextField phoneNumber;
-    @FXML private TextField email;
-    @FXML private TextField address;
+    @FXML private TextField companyName, contactPerson, phoneNumber, email, address;
     @FXML private ComboBox<Reputation> reputationComboBox;
 
     //Поля для удаления клиента
     @FXML private TextField deleteIdField;
 
     //Фильтры
-    @FXML private TextField idFilter;
-    @FXML private TextField companyNameFilter;
-    @FXML private TextField contactPersonFilter;
+    @FXML private TextField idFilter, companyNameFilter, contactPersonFilter;
     @FXML private ComboBox<String> reputationFilter;
     ObservableList<Client> observableClientList = FXCollections.observableArrayList();
 
+    @Autowired
+    private EmployeeService employeeService;
+    private Long userPassport;
+    private Role userRole;
 
-    public void initialize() {
+    public void initialize(Long userPassport) {
+        this.userPassport = userPassport;
+        this.userRole = employeeService.findEmployeeByPassportNumber(userPassport).get().getRole();
+
         ClientId.setCellValueFactory(new PropertyValueFactory<>("ClientId"));
         CompanyName.setCellValueFactory(new PropertyValueFactory<>("CompanyName"));
         ContactPerson.setCellValueFactory(new PropertyValueFactory<>("ContactPerson"));
@@ -108,11 +92,13 @@ public class ClientAppController {
         Email.setCellValueFactory(new PropertyValueFactory<>("Email"));
         Address.setCellValueFactory(new PropertyValueFactory<>("Address"));
         Reputation.setCellValueFactory(new PropertyValueFactory<>("Reputation"));
-        reputationComboBox.getItems().setAll(org.example.demotest.entities.Reputation.values());
 
-        companyNameFilter.setTextFormatter(createAlphaFilter());
-        contactPersonFilter.setTextFormatter(createAlphaFilter());
-        phoneNumber.setTextFormatter(createPhoneFilter());
+        if(userRole == org.example.demotest.entities.Role.ADMIN || userRole == org.example.demotest.entities.Role.MODERATOR) {
+            reputationComboBox.getItems().setAll(org.example.demotest.entities.Reputation.values());
+            companyNameFilter.setTextFormatter(createAlphaFilter());
+            contactPersonFilter.setTextFormatter(createAlphaFilter());
+            phoneNumber.setTextFormatter(createPhoneFilter());
+        }
 
         setupFilters();
         loadClients();
@@ -125,12 +111,14 @@ public class ClientAppController {
 
     @FXML
     private void handleCleanButton(){
-        companyName.setText(null);
-        contactPerson.setText(null);
-        phoneNumber.setText(null);
-        email.setText(null);
-        address.setText(null);
-        reputationComboBox.getSelectionModel().clearSelection();
+        if(userRole == org.example.demotest.entities.Role.ADMIN || userRole == org.example.demotest.entities.Role.MODERATOR) {
+            companyName.setText(null);
+            contactPerson.setText(null);
+            phoneNumber.setText(null);
+            email.setText(null);
+            address.setText(null);
+            reputationComboBox.getSelectionModel().clearSelection();
+        }
     }
 
     private void loadClients() {
@@ -151,19 +139,22 @@ public class ClientAppController {
     @FXML
     private void handleAddClient(ActionEvent event) {
         try {
-            ServiceRequestClient newClient = ServiceRequestClient.builder()
-                    .companyName(companyName.getText())
-                    .contactPerson(contactPerson.getText())
-                    .phoneNumber(phoneNumber.getText())
-                    .email(email.getText())
-                    .address(address.getText())
-                    .reputation(reputationComboBox.getValue())
-                    .build();
-            Client clientCreated = restTemplate.postForObject(url, newClient, Client.class);
-            if (clientCreated != null) {
-                observableClientList.add(clientCreated); // Используйте observableUserList
-                clientTable.setItems(observableClientList);
-                System.out.println("Пользователь добавлен: " + clientCreated);
+            if(userRole == org.example.demotest.entities.Role.ADMIN) {
+
+                ServiceRequestClient newClient = ServiceRequestClient.builder()
+                        .companyName(companyName.getText())
+                        .contactPerson(contactPerson.getText())
+                        .phoneNumber(phoneNumber.getText())
+                        .email(email.getText())
+                        .address(address.getText())
+                        .reputation(reputationComboBox.getValue())
+                        .build();
+                Client clientCreated = restTemplate.postForObject(url, newClient, Client.class);
+                if (clientCreated != null) {
+                    observableClientList.add(clientCreated); // Используйте observableUserList
+                    clientTable.setItems(observableClientList);
+                    System.out.println("Пользователь добавлен: " + clientCreated);
+                }
             }
         } catch (Exception e) {
             // Логирование ошибки и/или уведомление пользователя
@@ -203,20 +194,23 @@ public class ClientAppController {
 
     @FXML
     private void handleDeleteClient() {
-        Long idText = Long.valueOf(deleteIdField.getText());
+        if(userRole == org.example.demotest.entities.Role.ADMIN) {
 
-        if (idText != null) {
-            // Найдем пользователя по Id
-            Client client = clientService.findClientById(idText);
-            if (client != null) {
-                clientService.deleteClient(client.getClientId());
-                loadClients();
+            Long idText = Long.valueOf(deleteIdField.getText());
+
+            if (idText != null) {
+                // Найдем пользователя по Id
+                Client client = clientService.findClientById(idText);
+                if (client != null) {
+                    clientService.deleteClient(client.getClientId());
+                    loadClients();
+                } else {
+                    System.out.println("Клиент с указанным ID не найден");
+                }
             } else {
-                System.out.println("Клиент с указанным ID не найден");
+                // Обработка ситуации, если оба поля пусты
+                System.out.println("Введите ID для удаления клиента");
             }
-        } else {
-            // Обработка ситуации, если оба поля пусты
-            System.out.println("Введите ID для удаления клиента");
         }
     }
 
@@ -224,6 +218,6 @@ public class ClientAppController {
     public void handleBackButton(ActionEvent event) {
         Stage stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
         LoginManager loginManager = new LoginManager(stage, applicationContext);
-        loginManager.showMainView(String.valueOf(sessionID));
+        loginManager.showMainView(userPassport);
     }
 }
