+++ src/main/java/org/example/demotest/app_controllers/DepartmentAppController.java
@@ -10,16 +10,16 @@ import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
 import org.example.demotest.dto.ServiceRequestDepartment;
 import org.example.demotest.entities.Department;
+import org.example.demotest.entities.Role;
 import org.example.demotest.managers.LoginManager;
 import org.example.demotest.services.DepartmentService;
+import org.example.demotest.services.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.Arrays;
-
-import static org.example.demotest.app_controllers.LoginController.sessionID;
 
 @Controller
 public class DepartmentAppController {
@@ -45,8 +45,8 @@ public class DepartmentAppController {
             return null;
         });
     }
-    private RestTemplate restTemplate = new RestTemplate();
-    private String url = "http://localhost:8080/department-api/v1/departments";
+    private final RestTemplate restTemplate = new RestTemplate();
+    private final String url = "http://localhost:8080/department-api/v1/departments";
 
     //Таблица
     @FXML private TableView<Department> departmentTable;
@@ -69,8 +69,14 @@ public class DepartmentAppController {
     @FXML private TextField locationFilter;
     ObservableList<Department> observableDepartmentList = FXCollections.observableArrayList();
 
+    @Autowired
+    private EmployeeService employeeService;
+    private Long userPassport;
+    private Role userRole;
 
-    public void initialize() {
+    public void initialize(Long userPassport) {
+        this.userPassport = userPassport;
+        this.userRole = employeeService.findEmployeeByPassportNumber(userPassport).get().getRole();
         DepartmentId.setCellValueFactory(new PropertyValueFactory<>("DepartmentId"));
         DepartmentName.setCellValueFactory(new PropertyValueFactory<>("DepartmentName"));
         Location.setCellValueFactory(new PropertyValueFactory<>("Location"));
@@ -113,16 +119,19 @@ public class DepartmentAppController {
     @FXML
     private void handleAddDepartment(ActionEvent event) {
         try {
-            ServiceRequestDepartment newClient = ServiceRequestDepartment.builder()
-                    .departmentName(departmentName.getText())
-                    .location(locAtion.getText())
-                    .description(description.getText())
-                    .build();
-            Department departmentCreated = restTemplate.postForObject(url, newClient, Department.class);
-            if (departmentCreated != null) {
-                observableDepartmentList.add(departmentCreated); // Используйте observableUserList
-                departmentTable.setItems(observableDepartmentList);
-                System.out.println("Отдел добавлен: " + departmentCreated);
+            if(userRole == org.example.demotest.entities.Role.ADMIN) {
+
+                ServiceRequestDepartment newClient = ServiceRequestDepartment.builder()
+                        .departmentName(departmentName.getText())
+                        .location(locAtion.getText())
+                        .description(description.getText())
+                        .build();
+                Department departmentCreated = restTemplate.postForObject(url, newClient, Department.class);
+                if (departmentCreated != null) {
+                    observableDepartmentList.add(departmentCreated); // Используйте observableUserList
+                    departmentTable.setItems(observableDepartmentList);
+                    System.out.println("Отдел добавлен: " + departmentCreated);
+                }
             }
         } catch (Exception e) {
             // Логирование ошибки и/или уведомление пользователя
@@ -160,20 +169,23 @@ public class DepartmentAppController {
 
     @FXML
     private void handleDeleteDepartment() {
-        Long idText = Long.valueOf(deleteIdField.getText());
+        if (userRole == org.example.demotest.entities.Role.ADMIN) {
 
-        if (idText != null) {
-            // Найдем отдел по Id
-            Department department = departmentService.findDepartmentById(idText);
-            if (department != null) {
-                departmentService.deleteDepartment(department.getDepartmentId());
-                loadDepartments();
+            Long idText = Long.valueOf(deleteIdField.getText());
+
+            if (idText != null) {
+                // Найдем отдел по Id
+                Department department = departmentService.findDepartmentById(idText);
+                if (department != null) {
+                    departmentService.deleteDepartment(department.getDepartmentId());
+                    loadDepartments();
+                } else {
+                    System.out.println("Клиент с указанным ID не найден");
+                }
             } else {
-                System.out.println("Клиент с указанным ID не найден");
+                // Обработка ситуации, если оба поля пусты
+                System.out.println("Введите ID для удаления клиента");
             }
-        } else {
-            // Обработка ситуации, если оба поля пусты
-            System.out.println("Введите ID для удаления клиента");
         }
     }
 
@@ -181,6 +193,6 @@ public class DepartmentAppController {
     public void handleBackButton(ActionEvent event) {
         Stage stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
         LoginManager loginManager = new LoginManager(stage, applicationContext);
-        loginManager.showMainView(String.valueOf(sessionID));
+        loginManager.showMainView(userPassport);
     }
 }
